requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# Explainability
shap>=0.43.0
lime>=0.2.0.1

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Dashboard
streamlit>=1.28.0

# Configuration Management
pyyaml>=6.0
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.5.0

# Logging
loguru>=0.7.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Progress Bars
tqdm>=4.66.0

# HTTP requests (for dataset download)
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Model Serialization
joblib>=1.3.0

# Utilities
openpyxl>=3.1.0  # Excel support
